# api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: TrackBot App API
  description: |
    Comprehensive API for the TrackBot Application helping in logging workouts, exercises and nutrition using AI.
    
    This API provides endpoints for:
    - User management
    - Workout tracking and management
    - Exercise logging (cardio and weight training)
    - AI conversation and messaging
    
    All timestamps are in ISO 8601 format (UTC).
  version: 1.0.0
  contact:
    name: Utkarsh Khandelwal
    email: utkarshkhandelwal2011@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.trackbotapp.com/v1
    description: Production server
...
paths:
...
  /users/{userId}/workouts:
    get:
      summary: List workouts for a user
      description: Retrieve all workouts for a specific user with optional date filtering
      operationId: listWorkouts
      tags: [Workouts]
      parameters:
        - $ref: './components/parameters/common.yaml#/UserIdPath'
        - $ref: './components/parameters/common.yaml#/YearQuery'
        - $ref: './components/parameters/common.yaml#/MonthQuery'
        - $ref: './components/parameters/common.yaml#/DayQuery'
      responses:
        '200':
          description: List of workouts retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/workout.yaml#/components/schemas/ListWorkoutsResponse'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    post:
      summary: Create a new workout
      description: Create a new workout session for a user
      operationId: createWorkout
      tags: [Workouts]
      parameters:
        - $ref: './components/parameters/common.yaml#/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/workout.yaml#/components/schemas/CreateWorkoutRequest'
      responses:
        '201':
          description: Workout created successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/workout.yaml#/components/schemas/CreateWorkoutResponse'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
  /workouts/{workoutId}:
    get:
      summary: Get workout by ID
      description: Retrieve a specific workout by its ID
      operationId: getWorkoutById
      tags: [Workouts]
      parameters:
        - $ref: './components/parameters/common.yaml#/WorkoutIdPath'
      responses:
        '200':
          description: Workout details retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/workout.yaml#/components/schemas/Workout'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    put:
      summary: Update workout
      description: Update an existing workout
      operationId: updateWorkout
      tags: [Workouts]
      parameters:
        - $ref: './components/parameters/common.yaml#/WorkoutIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/workout.yaml#/components/schemas/UpdateWorkoutRequest'
      responses:
        '200':
          description: Workout updated successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/workout.yaml#/components/schemas/Workout'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    delete:
      summary: Delete workout
      description: Delete a workout and all associated exercises
      operationId: deleteWorkout
      tags: [Workouts]
      parameters:
        - $ref: './components/parameters/common.yaml#/WorkoutIdPath'
      responses:
        '204':
          $ref: './components/responses/errors.yaml#/components/responses/NoContent'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'

  # Exercise endpoints
...
components:
  schemas:
...
    Workout:
      $ref: './components/schemas/workout.yaml#/components/schemas/Workout'
...
    CreateWorkoutRequest:
      $ref: './components/schemas/workout.yaml#/components/schemas/CreateWorkoutRequest'
    UpdateWorkoutRequest:
      $ref: './components/schemas/workout.yaml#/components/schemas/UpdateWorkoutRequest'
    CreateWorkoutResponse:
      $ref: './components/schemas/workout.yaml#/components/schemas/CreateWorkoutResponse'
    ListWorkoutsResponse:
      $ref: './components/schemas/workout.yaml#/components/schemas/ListWorkoutsResponse'
...
  responses:
    BadRequest:
      $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
...
    NotFound:
      $ref: './components/responses/errors.yaml#/components/responses/NotFound'
...
    InternalServerError:
      $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    NoContent:
      $ref: './components/responses/errors.yaml#/components/responses/NoContent'