# backend/services/workoutAppServices/internal/generated/model_create_workout_response.go
/*
TrackBot App API

Comprehensive API for the TrackBot Application helping in logging workouts, exercises and nutrition using AI.  This API provides endpoints for: - User management - Workout tracking and management - Exercise logging (cardio and weight training) - AI conversation and messaging  All timestamps are in ISO 8601 format (UTC). 

API version: 1.0.0
Contact: utkarshkhandelwal2011@gmail.com
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package api_models

import (
	"encoding/json"
	"bytes"
	"fmt"
)

// checks if the CreateWorkoutResponse type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &CreateWorkoutResponse{}

// CreateWorkoutResponse Response returned after creating a workout.
type CreateWorkoutResponse struct {
	// ID of the created workout.
	Id int64 `json:"id"`
}

type _CreateWorkoutResponse CreateWorkoutResponse

// NewCreateWorkoutResponse instantiates a new CreateWorkoutResponse object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewCreateWorkoutResponse(id int64) *CreateWorkoutResponse {
	this := CreateWorkoutResponse{}
	this.Id = id
	return &this
}

// NewCreateWorkoutResponseWithDefaults instantiates a new CreateWorkoutResponse object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewCreateWorkoutResponseWithDefaults() *CreateWorkoutResponse {
	this := CreateWorkoutResponse{}
	return &this
}

// GetId returns the Id field value
func (o *CreateWorkoutResponse) GetId() int64 {
	if o == nil {
		var ret int64
		return ret
	}

	return o.Id
}

// GetIdOk returns a tuple with the Id field value
// and a boolean to check if the value has been set.
func (o *CreateWorkoutResponse) GetIdOk() (*int64, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Id, true
}

// SetId sets field value
func (o *CreateWorkoutResponse) SetId(v int64) {
	o.Id = v
}

func (o CreateWorkoutResponse) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o CreateWorkoutResponse) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["id"] = o.Id
	return toSerialize, nil
}

func (o *CreateWorkoutResponse) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"id",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varCreateWorkoutResponse := _CreateWorkoutResponse{}

	decoder := json.NewDecoder(bytes.NewReader(data))
	decoder.DisallowUnknownFields()
	err = decoder.Decode(&varCreateWorkoutResponse)

	if err != nil {
		return err
	}

	*o = CreateWorkoutResponse(varCreateWorkoutResponse)

	return err
}

type NullableCreateWorkoutResponse struct {
	value *CreateWorkoutResponse
	isSet bool
}

func (v NullableCreateWorkoutResponse) Get() *CreateWorkoutResponse {
	return v.value
}

func (v *NullableCreateWorkoutResponse) Set(val *CreateWorkoutResponse) {
	v.value = val
	v.isSet = true
}

func (v NullableCreateWorkoutResponse) IsSet() bool {
	return v.isSet
}

func (v *NullableCreateWorkoutResponse) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableCreateWorkoutResponse(val *CreateWorkoutResponse) *NullableCreateWorkoutResponse {
	return &NullableCreateWorkoutResponse{value: val, isSet: true}
}

func (v NullableCreateWorkoutResponse) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableCreateWorkoutResponse) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}