# backend/services/workoutAppServices/internal/generated/model_workout.go
/*
TrackBot App API

Comprehensive API for the TrackBot Application helping in logging workouts, exercises and nutrition using AI.  This API provides endpoints for: - User management - Workout tracking and management - Exercise logging (cardio and weight training) - AI conversation and messaging  All timestamps are in ISO 8601 format (UTC). 

API version: 1.0.0
Contact: utkarshkhandelwal2011@gmail.com
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package api_models

import (
	"encoding/json"
	"time"
	"bytes"
	"fmt"
)

// checks if the Workout type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &Workout{}

// Workout Represents a workout session.
type Workout struct {
	// Unique identifier for the workout.
	Id int64 `json:"id"`
	// ID of the user who owns the workout.
	UserId int64 `json:"user_id"`
	// When the workout was created.
	CreatedAt time.Time `json:"created_at"`
	// When the workout was last updated.
	UpdatedAt time.Time `json:"updated_at"`
}

type _Workout Workout

// NewWorkout instantiates a new Workout object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewWorkout(id int64, userId int64, createdAt time.Time, updatedAt time.Time) *Workout {
	this := Workout{}
	this.Id = id
	this.UserId = userId
	this.CreatedAt = createdAt
	this.UpdatedAt = updatedAt
	return &this
}

// NewWorkoutWithDefaults instantiates a new Workout object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewWorkoutWithDefaults() *Workout {
	this := Workout{}
	return &this
}

// GetId returns the Id field value
func (o *Workout) GetId() int64 {
	if o == nil {
		var ret int64
		return ret
	}

	return o.Id
}

// GetIdOk returns a tuple with the Id field value
// and a boolean to check if the value has been set.
func (o *Workout) GetIdOk() (*int64, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Id, true
}

// SetId sets field value
func (o *Workout) SetId(v int64) {
	o.Id = v
}

// GetUserId returns the UserId field value
func (o *Workout) GetUserId() int64 {
	if o == nil {
		var ret int64
		return ret
	}

	return o.UserId
}

// GetUserIdOk returns a tuple with the UserId field value
// and a boolean to check if the value has been set.
func (o *Workout) GetUserIdOk() (*int64, bool) {
	if o == nil {
		return nil, false
	}
	return &o.UserId, true
}

// SetUserId sets field value
func (o *Workout) SetUserId(v int64) {
	o.UserId = v
}

// GetCreatedAt returns the CreatedAt field value
func (o *Workout) GetCreatedAt() time.Time {
	if o == nil {
		var ret time.Time
		return ret
	}

	return o.CreatedAt
}

// GetCreatedAtOk returns a tuple with the CreatedAt field value
// and a boolean to check if the value has been set.
func (o *Workout) GetCreatedAtOk() (*time.Time, bool) {
	if o == nil {
		return nil, false
	}
	return &o.CreatedAt, true
}

// SetCreatedAt sets field value
func (o *Workout) SetCreatedAt(v time.Time) {
	o.CreatedAt = v
}

// GetUpdatedAt returns the UpdatedAt field value
func (o *Workout) GetUpdatedAt() time.Time {
	if o == nil {
		var ret time.Time
		return ret
	}

	return o.UpdatedAt
}

// GetUpdatedAtOk returns a tuple with the UpdatedAt field value
// and a boolean to check if the value has been set.
func (o *Workout) GetUpdatedAtOk() (*time.Time, bool) {
	if o == nil {
		return nil, false
	}
	return &o.UpdatedAt, true
}

// SetUpdatedAt sets field value
func (o *Workout) SetUpdatedAt(v time.Time) {
	o.UpdatedAt = v
}

func (o Workout) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o Workout) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["id"] = o.Id
	toSerialize["user_id"] = o.UserId
	toSerialize["created_at"] = o.CreatedAt
	toSerialize["updated_at"] = o.UpdatedAt
	return toSerialize, nil
}

func (o *Workout) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"id",
		"user_id",
		"created_at",
		"updated_at",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varWorkout := _Workout{}

	decoder := json.NewDecoder(bytes.NewReader(data))
	decoder.DisallowUnknownFields()
	err = decoder.Decode(&varWorkout)

	if err != nil {
		return err
	}

	*o = Workout(varWorkout)

	return err
}

type NullableWorkout struct {
	value *Workout
	isSet bool
}

func (v NullableWorkout) Get() *Workout {
	return v.value
}

func (v *NullableWorkout) Set(val *Workout) {
	v.value = val
	v.isSet = true
}

func (v NullableWorkout) IsSet() bool {
	return v.isSet
}

func (v *NullableWorkout) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableWorkout(val *Workout) *NullableWorkout {
	return &NullableWorkout{value: val, isSet: true}
}

func (v NullableWorkout) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableWorkout) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}